+++ src/Adapter/PlatformServices.Desktop/Services/DesktopTestSourceHost.cs
@@ -14,8 +14,6 @@ namespace Microsoft.VisualStudio.TestPlatform.MSTestAdapter.PlatformServices
     using Microsoft.VisualStudio.TestPlatform.ObjectModel.Adapter;
     using Microsoft.VisualStudio.TestPlatform.ObjectModel.Utilities;
 
-#pragma warning disable SA1649 // SA1649FileNameMustMatchTypeName
-
     /// <summary>
     /// A host that loads the test source.This can be in isolation for desktop using an AppDomain or just loading the source in the current context.
     /// </summary>
