+++ src/Adapter/PlatformServices.Desktop/Utilities/DesktopAssemblyUtility.cs
@@ -13,8 +13,6 @@ namespace Microsoft.VisualStudio.TestPlatform.MSTestAdapter.PlatformServices.Uti
     using Microsoft.VisualStudio.TestPlatform.MSTestAdapter.PlatformServices.Deployment;
     using Microsoft.VisualStudio.TestPlatform.ObjectModel;
 
-#pragma warning disable SA1649 // File name must match first type name
-
     /// <summary>
     /// Utility for assembly specific functionality.
     /// </summary>
