+++ src/Adapter/PlatformServices.Shared/netstandard1.0/Services/ns10FileOperations.cs
