+++ src/Adapter/PlatformServices.Shared/netstandard1.3/Utilities/ns13DeploymentUtilityBase.cs
@@ -19,8 +19,6 @@ namespace Microsoft.VisualStudio.TestPlatform.MSTestAdapter.PlatformServices.Uti
     using Microsoft.VisualStudio.TestPlatform.ObjectModel.Adapter;
     using Microsoft.VisualStudio.TestPlatform.ObjectModel.Logging;
 
-#pragma warning disable SA1649 // File name must match first type name
-
     internal abstract class DeploymentUtilityBase
     {
         protected const string TestAssemblyConfigFileExtension = ".config";
