+++ test/UnitTests/PlatformServices.Shared.Unit.Tests/netstandard1.0/Deployment/ns10TestRunDirectoriesTests.cs
@@ -16,8 +16,6 @@ namespace MSTestAdapter.PlatformServices.Tests.Deployment
     using System.IO;
     using Microsoft.VisualStudio.TestPlatform.MSTestAdapter.PlatformServices.Deployment;
 
-#pragma warning disable SA1649 // File name must match first type name
-
     [TestClass]
     public class TestRunDirectoriesTests
     {
