+++ test/UnitTests/PlatformServices.Shared.Unit.Tests/netstandard1.0/ns10FileOperationsTests.cs
@@ -19,7 +19,6 @@ namespace MSTestAdapter.PlatformServices.Tests.Services
     using System.Reflection;
     using Microsoft.VisualStudio.TestPlatform.MSTestAdapter.PlatformServices;
     using MSTestAdapter.TestUtilities;
-#pragma warning disable SA1649 // SA1649FileNameMustMatchTypeName
 
     [TestClass]
     public class FileOperationsTests
