+++ test/UnitTests/PlatformServices.Shared.Unit.Tests/netstandard1.0/ns10SettingsProviderTests.cs
@@ -16,8 +16,6 @@ namespace MSTestAdapter.PlatformServices.Tests.Services
     using Microsoft.VisualStudio.TestPlatform.MSTestAdapter.PlatformServices;
     using Moq;
 
-#pragma warning disable SA1649 // SA1649FileNameMustMatchTypeName
-
     [TestClass]
     public class SettingsProviderTests
     {
