+++ test/UnitTests/PlatformServices.Shared.Unit.Tests/netstandard1.0/ns10TestSourceTests.cs
@@ -20,8 +20,6 @@ namespace MSTestAdapter.PlatformServices.Tests.Services
 
     using Microsoft.VisualStudio.TestPlatform.MSTestAdapter.PlatformServices;
 
-#pragma warning disable SA1649 // SA1649FileNameMustMatchTypeName
-
     [TestClass]
     public class TestSourceTests
     {
